Move `PdfNavigatorFragment` into a new `readium-adapter-pdfium-navigator` module
