+++ readium/adapters/pdfium/pdfium-document/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="org.readium.adapters.pdfium.document" />
