+++ readium/adapters/pdfium/pdfium-navigator/build.gradle
@@ -48,7 +48,7 @@ afterEvaluate {
                 from components.release
 
                 groupId = 'com.github.readium'
-                artifactId = 'readium-navigator-pdf'
+                artifactId = 'readium-adapter-pdfium-navigator'
 
                 artifact sourcesJar
                 artifact javadocsJar
