+++ readium/navigator/build.gradle
@@ -54,5 +54,4 @@ dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
 
     api project(':readium:navigator:navigator-core')
-    api project(':readium:navigator:navigator-pdf')
 }
