--- readium/navigator/navigator-pdf/src/main/AndroidManifest.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.readium.r2.navigator.pdf">
-
-    <application>
-        <activity
-            android:name=".R2PdfActivity"
-            android:theme="@style/AppTheme" />
-    </application>
-</manifest>
