+++ test-app/build.gradle
@@ -83,6 +83,8 @@ dependencies {
     implementation project(':readium:opds')
     implementation project(':readium:lcp')
 
+    implementation project(':readium:adapters:pdfium')
+
     implementation 'androidx.core:core-ktx:1.6.0'
     implementation "androidx.activity:activity-ktx:1.3.1"
     implementation "androidx.appcompat:appcompat:1.3.1"
